+++ docs/CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 0.1-dev (unreleased)
 --------------------
 
+- Improve amount parsing for prepaid accounts.
+  [davidjb]
 - Use improved logging formatting for module.
   [davidjb]
 - Add send credit script to allow scripting of sending credit.